FIX: writeComment delete
