--- src/main/java/com/knu/community/board/repository/WriteBoardRepository.java
@@ -1,8 +1,0 @@
-package com.knu.community.board.repository;
-
-import com.knu.community.board.domain.WriteBoard;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WriteBoardRepository extends JpaRepository<WriteBoard,Long> {
-
-}
