+++ src/main/java/com/knu/community/comment/service/CommentService.java
@@ -3,17 +3,13 @@ package com.knu.community.comment.service;
 import com.knu.community.board.domain.Board;
 import com.knu.community.board.repository.BoardRepository;
 import com.knu.community.comment.domain.Comment;
-import com.knu.community.comment.domain.WriteComment;
 import com.knu.community.comment.dto.CommentForm;
 import com.knu.community.comment.repository.CommentRepository;
-import com.knu.community.comment.repository.WriteCommentRepository;
 import com.knu.community.error.NotFoundException;
 import com.knu.community.member.domain.Member;
 import com.knu.community.member.repository.MemberRepository;
-import com.knu.community.reply.domain.Reply;
 import com.knu.community.reply.repository.ReplyRepository;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,10 +19,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
     private final CommentRepository commentRepository;
-    private final ReplyRepository replyRepository;
     private final BoardRepository boardRepository;
     private final MemberRepository memberRepository;
-    private final WriteCommentRepository writeCommentRepository;
 
     @Transactional
     public Comment writeComment(Long boardId,Long userId, CommentForm commentForm){
@@ -34,11 +28,8 @@ public class CommentService {
             new NotFoundException("Member가 존재하지 않습니다."));
         Board board = boardRepository.findById(boardId).orElseThrow(()->
             new NotFoundException("Comment를 달 수 있는 Board가 존재하지 않습니다."));
-        Comment comment = createComment(board,commentForm);
+        Comment comment = createComment(member,board,commentForm);
 
-        WriteComment writeComment=new WriteComment(member,comment);
-
-        writeCommentRepository.save(writeComment);
 
         return comment;
     }
@@ -60,8 +51,8 @@ public class CommentService {
     public void deleteComment(Long commentId){boardRepository.deleteById(commentId);}
 
     @Transactional
-    public Comment createComment(Board board,CommentForm commentForm){
-        Comment comment = new Comment(board,commentForm);
+    public Comment createComment(Member member,Board board,CommentForm commentForm){
+        Comment comment = new Comment(member,board,commentForm);
         commentRepository.save(comment);
 
         return comment;
